Change to long

+++ features/transaction/checkout/src/main/java/com/tokopedia/checkout/view/view/cartlist/CartListPresenter.java
@@ -339,7 +339,7 @@ public class CartListPresenter implements ICartListPresenter {
         }
 
         view.renderDetailInfoSubTotal(String.valueOf(totalAllCartItemQty),
-                CurrencyFormatUtil.convertPriceValueToIdrFormat(((int) subtotalPrice), true));
+                CurrencyFormatUtil.convertPriceValueToIdrFormat(((long) subtotalPrice), true));
     }
 
     @Override
